requirements.txt
Flask==2.0.3
Flask-WTF==1.0.0
PyYAML==6.0
Markdown==3.3.7
Frozen-Flask==0.18

